+++ spring-web-examples/spring-web-example-realtime/src/main/java/io/springperf/example/realtime/RealtimeApplication.java
@@ -1,0 +1,12 @@
+package io.springperf.example.realtime;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RealtimeApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(RealtimeApplication.class, args);
+    }
+}