+++ spring-web-examples/spring-web-example-rest/src/main/java/io/springperf/example/rest/model/User.java
@@ -1,0 +1,15 @@
+package io.springperf.example.rest.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class User {
+    private Long id;
+    private String name;
+    private String email;
+    private Integer age;
+}